+++ clang-tools-extra/test/clang-tidy/checkers/readability/if-block-size.cpp
@@ -22,6 +22,31 @@ void should_warn(){
                 + 19
                 + 20;
     } //          21
+
+
+    if (true){ // 1
+        int sum = 2;
+    } else { //   3      1
+// CHECK-MESSAGES: :[[@LINE-1]]:12: warning: else block spans 21 lines of code, which exceeds the threshold of 20 lines [readability-if-block-size]
+        int sum = 5   // 3
+                + 6   // 4
+                + 7   // 5
+                + 8   // 6
+                + 9   // 7
+                + 10  // 8
+                + 11  // 9
+                + 12  // 10
+                + 13  // 11
+                + 14  // 12
+                + 15  // 13
+                + 16  // 14
+                + 17  // 15
+                + 18  // 16
+                + 19  // 17
+                + 20  // 18
+                + 21  // 19
+                + 22; // 20
+    } //          23     21
 }
 
 void should_not_warn(){