+++ clang-tools-extra/docs/ReleaseNotes.md
@@ -119,6 +119,11 @@ infrastructure are described first, followed by tool-specific sections.
   Finds uses of `#pragma comment` and, for `lib` or `linker` comments, suggests
   using the build system for improved portability.
 
+- New {doc}`readability-if-block-size
+  <clang-tidy/checks/readability/if-block-size>` check.
+
+  FIXME: Write a short description.
+
 - New {doc}`readability-redundant-zero-initializer
   <clang-tidy/checks/readability/redundant-zero-initializer>` check.